Fold tests into redcard.js, bucket by gen, drop issue ref
Per review on #12070: removed the standalone test file, moved the
Gen 6 and Gen 5 cases into the existing redcard.js inside nested
describe('[Gen N]', ...) blocks, and dropped the #8504 reference.
